Move Performance function to main Testbed file and remove some unneeded usings
